feat: Add support for hiding student names [PT-184362179]
This change hides the student names in the details, usage and "wide" log report if the signed request params from the server contains a value for "hide_names".
