+++ query-creator/create-query/app.js
@@ -89,6 +89,9 @@ const learnersReport = async (params, body, tokenServiceEnv, debugSQL, reportSer
   const { query, portal_url, learnersApiUrl, user } = json;
   let { email } = user;
 
+  // currently this value is a string in the params generated by the portal but this ensures it it changes to a boolean this won't break
+  const hideNames = query.hide_names === "true" || query.hide_names === true;
+
   const portalUrl = trimTrailingSlash(portal_url);
 
   const authDomain = portalToAuthDomainMap[portalUrl] || portalUrl;
@@ -123,7 +126,7 @@ const learnersReport = async (params, body, tokenServiceEnv, debugSQL, reportSer
     }
 
     // generate the sql for the query
-    const sql = aws.generateSQL(runnableInfo, usageReport, authDomain, reportServiceSource);
+    const sql = aws.generateSQL(runnableInfo, usageReport, authDomain, reportServiceSource, hideNames);
 
     if (debugSQL) {
       sqlOutput.push(sql);
@@ -136,8 +139,8 @@ const learnersReport = async (params, body, tokenServiceEnv, debugSQL, reportSer
   const doLearnerLogReporting= async () => {
     // generate the sql for the query
     const sql = narrowLearners
-      ? aws.generateNarrowLogSQL(queryIdsPerRunnable, authDomain, reportServiceSource)
-      : aws.generateLearnerLogSQL(queryIdsPerRunnable, authDomain, reportServiceSource);
+      ? aws.generateNarrowLogSQL(queryIdsPerRunnable, authDomain, reportServiceSource) // hideNames not needed here as no learner info is output
+      : aws.generateLearnerLogSQL(queryIdsPerRunnable, authDomain, reportServiceSource, hideNames);
 
     if (debugSQL) {
       sqlOutput.push(sql);
